LibWeb: Scale line height by 1.2 for `line-height: normal`
Other browsers such as Chrome and Firefox also apply the 1.2 scale to
the font family's line height for the `normal` keyword. The spec
mentions a "line gap" value per font, which is already included in
`FontMetrics::line_height()`.

This also requires updating the expectations for a lot of layout tests.
Fortunately, this means that most of these now result in pixel-perfect
line heights when compared to other browsers :^)