CI: Conditionally fall back to GitHub cache for macOS builds
The Blacksmith cache is only reachable by Blacksmith runners, so we need
to explicitly use a different cache action for GitHub runners and
self-hosted runners.

This should be a temporary solution since a new version of the cache
action should be released soon that supports a transparent fallback to
the GitHub cache.