LibWeb: Store both focusable type and anchor in Document
Instead of just keeping track of the focusable anchor, we now accept
an Optional<FocusableArea> so we keep track of the type as well.

Note that we do not actually reuse FocusableArea for storage in
Document, since we have style logic that depends on a mutable
GC::Ptr<Node> even if there is no focused area.